refactor: move campaign selector logic into CampaignSelector dir
